Merge rpcs3 Changes

+++ rpcs3/Gui/GameViewer.cpp
@@ -6,6 +6,8 @@
 #include "Loader/PSF.h"
 #include "SettingsDialog.h"
 
+#include <algorithm>
+
 static const std::string m_class_name = "GameViewer";
 
 // Auxiliary classes
@@ -233,8 +235,11 @@ void GameViewer::RemoveGame(wxCommandEvent& event)
 {
 	long i = GetFirstSelected();
 	if (i < 0) return;
-	
-	fs::remove_all(vfs::get(m_path) + this->GetItemText(i, 6).ToStdString());
+
+	if (wxMessageBox("Permanently delete game files?", "Confirm Delete", wxYES_NO | wxNO_DEFAULT) == wxYES)
+	{
+		fs::remove_all(vfs::get(m_path) + this->GetItemText(i, 6).ToStdString());
+	}
 
 	Refresh();
 }
